Update leds-buttons.md
The LED code example incorrectly takes the line HIGH with the comment saying the LED will be on, etc. Probably hasn't been noticed because the code will blink the led - only it goes 'off on off', rather than 'on off on'.

(The BUTTONS code example is correct)

This correction fixes that.